proxycfg: move test helpers to proxycfgtest
