+++ agent/xds/clusters_test.go
@@ -10,12 +10,13 @@ import (
 
 	envoy "github.com/envoyproxy/go-control-plane/envoy/api/v2"
 	"github.com/golang/protobuf/ptypes/wrappers"
+	"github.com/stretchr/testify/require"
+
 	"github.com/hashicorp/consul/agent/proxycfg"
+	"github.com/hashicorp/consul/agent/proxycfg/proxycfgtest"
 	"github.com/hashicorp/consul/agent/structs"
 	"github.com/hashicorp/consul/agent/xds/proxysupport"
 	"github.com/hashicorp/consul/sdk/testutil"
-	testinf "github.com/mitchellh/go-testing-interface"
-	"github.com/stretchr/testify/require"
 )
 
 func TestClustersFromSnapshot(t *testing.T) {
@@ -25,7 +26,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 
 	tests := []struct {
 		name   string
-		create func(t testinf.T) *proxycfg.ConfigSnapshot
+		create func(t *testing.T) *proxycfg.ConfigSnapshot
 		// Setup is called before the test starts. It is passed the snapshot from
 		// create func and is allowed to modify it in any way to setup the
 		// test input.
@@ -34,12 +35,12 @@ func TestClustersFromSnapshot(t *testing.T) {
 	}{
 		{
 			name:   "defaults",
-			create: proxycfg.TestConfigSnapshot,
+			create: proxycfgtest.TestConfigSnapshot,
 			setup:  nil, // Default snapshot
 		},
 		{
 			name:   "custom-local-app",
-			create: proxycfg.TestConfigSnapshot,
+			create: proxycfgtest.TestConfigSnapshot,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.Proxy.Config["envoy_local_cluster_json"] =
 					customAppClusterJSON(t, customClusterJSONOptions{
@@ -50,7 +51,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:               "custom-local-app-typed",
-			create:             proxycfg.TestConfigSnapshot,
+			create:             proxycfgtest.TestConfigSnapshot,
 			overrideGoldenName: "custom-local-app",
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.Proxy.Config["envoy_local_cluster_json"] =
@@ -62,7 +63,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "custom-upstream",
-			create: proxycfg.TestConfigSnapshot,
+			create: proxycfgtest.TestConfigSnapshot,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.Proxy.Upstreams[0].Config["envoy_cluster_json"] =
 					customAppClusterJSON(t, customClusterJSONOptions{
@@ -73,7 +74,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "custom-upstream-default-chain",
-			create: proxycfg.TestConfigSnapshotDiscoveryChainDefault,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChainDefault,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.Proxy.Upstreams[0].Config["envoy_cluster_json"] =
 					customAppClusterJSON(t, customClusterJSONOptions{
@@ -84,7 +85,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:               "custom-upstream-typed",
-			create:             proxycfg.TestConfigSnapshot,
+			create:             proxycfgtest.TestConfigSnapshot,
 			overrideGoldenName: "custom-upstream",
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.Proxy.Upstreams[0].Config["envoy_cluster_json"] =
@@ -96,7 +97,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:               "custom-upstream-ignores-tls",
-			create:             proxycfg.TestConfigSnapshot,
+			create:             proxycfgtest.TestConfigSnapshot,
 			overrideGoldenName: "custom-upstream", // should be the same
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.Proxy.Upstreams[0].Config["envoy_cluster_json"] =
@@ -110,7 +111,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "custom-timeouts",
-			create: proxycfg.TestConfigSnapshot,
+			create: proxycfgtest.TestConfigSnapshot,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.Proxy.Config["local_connect_timeout_ms"] = 1234
 				snap.Proxy.Upstreams[0].Config["connect_timeout_ms"] = 2345
@@ -118,7 +119,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "custom-limits-max-connections-only",
-			create: proxycfg.TestConfigSnapshot,
+			create: proxycfgtest.TestConfigSnapshot,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				for i := range snap.Proxy.Upstreams {
 					// We check if Config is nil because the prepared_query upstream is
@@ -137,7 +138,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "custom-limits-set-to-zero",
-			create: proxycfg.TestConfigSnapshot,
+			create: proxycfgtest.TestConfigSnapshot,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				for i := range snap.Proxy.Upstreams {
 					if snap.Proxy.Upstreams[i].Config == nil {
@@ -154,7 +155,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "custom-limits",
-			create: proxycfg.TestConfigSnapshot,
+			create: proxycfgtest.TestConfigSnapshot,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				for i := range snap.Proxy.Upstreams {
 					if snap.Proxy.Upstreams[i].Config == nil {
@@ -171,81 +172,81 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "connect-proxy-with-chain",
-			create: proxycfg.TestConfigSnapshotDiscoveryChain,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChain,
 			setup:  nil,
 		},
 		{
 			name:   "connect-proxy-with-chain-external-sni",
-			create: proxycfg.TestConfigSnapshotDiscoveryChainExternalSNI,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChainExternalSNI,
 			setup:  nil,
 		},
 		{
 			name:   "connect-proxy-with-chain-and-overrides",
-			create: proxycfg.TestConfigSnapshotDiscoveryChainWithOverrides,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChainWithOverrides,
 			setup:  nil,
 		},
 		{
 			name:   "connect-proxy-with-chain-and-failover",
-			create: proxycfg.TestConfigSnapshotDiscoveryChainWithFailover,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChainWithFailover,
 			setup:  nil,
 		},
 		{
 			name:   "connect-proxy-with-tcp-chain-failover-through-remote-gateway",
-			create: proxycfg.TestConfigSnapshotDiscoveryChainWithFailoverThroughRemoteGateway,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChainWithFailoverThroughRemoteGateway,
 			setup:  nil,
 		},
 		{
 			name:   "connect-proxy-with-tcp-chain-failover-through-remote-gateway-triggered",
-			create: proxycfg.TestConfigSnapshotDiscoveryChainWithFailoverThroughRemoteGatewayTriggered,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChainWithFailoverThroughRemoteGatewayTriggered,
 			setup:  nil,
 		},
 		{
 			name:   "connect-proxy-with-tcp-chain-double-failover-through-remote-gateway",
-			create: proxycfg.TestConfigSnapshotDiscoveryChainWithDoubleFailoverThroughRemoteGateway,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChainWithDoubleFailoverThroughRemoteGateway,
 			setup:  nil,
 		},
 		{
 			name:   "connect-proxy-with-tcp-chain-double-failover-through-remote-gateway-triggered",
-			create: proxycfg.TestConfigSnapshotDiscoveryChainWithDoubleFailoverThroughRemoteGatewayTriggered,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChainWithDoubleFailoverThroughRemoteGatewayTriggered,
 			setup:  nil,
 		},
 		{
 			name:   "connect-proxy-with-tcp-chain-failover-through-local-gateway",
-			create: proxycfg.TestConfigSnapshotDiscoveryChainWithFailoverThroughLocalGateway,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChainWithFailoverThroughLocalGateway,
 			setup:  nil,
 		},
 		{
 			name:   "connect-proxy-with-tcp-chain-failover-through-local-gateway-triggered",
-			create: proxycfg.TestConfigSnapshotDiscoveryChainWithFailoverThroughLocalGatewayTriggered,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChainWithFailoverThroughLocalGatewayTriggered,
 			setup:  nil,
 		},
 		{
 			name:   "connect-proxy-with-tcp-chain-double-failover-through-local-gateway",
-			create: proxycfg.TestConfigSnapshotDiscoveryChainWithDoubleFailoverThroughLocalGateway,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChainWithDoubleFailoverThroughLocalGateway,
 			setup:  nil,
 		},
 		{
 			name:   "connect-proxy-with-tcp-chain-double-failover-through-local-gateway-triggered",
-			create: proxycfg.TestConfigSnapshotDiscoveryChainWithDoubleFailoverThroughLocalGatewayTriggered,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChainWithDoubleFailoverThroughLocalGatewayTriggered,
 			setup:  nil,
 		},
 		{
 			name:   "splitter-with-resolver-redirect",
-			create: proxycfg.TestConfigSnapshotDiscoveryChain_SplitterWithResolverRedirectMultiDC,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChain_SplitterWithResolverRedirectMultiDC,
 			setup:  nil,
 		},
 		{
 			name:   "connect-proxy-lb-in-resolver",
-			create: proxycfg.TestConfigSnapshotDiscoveryChainWithLB,
+			create: proxycfgtest.TestConfigSnapshotDiscoveryChainWithLB,
 			setup:  nil,
 		},
 		{
 			name:   "expose-paths-local-app-paths",
-			create: proxycfg.TestConfigSnapshotExposeConfig,
+			create: proxycfgtest.TestConfigSnapshotExposeConfig,
 		},
 		{
 			name:   "expose-paths-new-cluster-http2",
-			create: proxycfg.TestConfigSnapshotExposeConfig,
+			create: proxycfgtest.TestConfigSnapshotExposeConfig,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.Proxy.Expose.Paths[1] = structs.ExposePath{
 					LocalPathPort: 9090,
@@ -257,26 +258,26 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "expose-paths-grpc-new-cluster-http1",
-			create: proxycfg.TestConfigSnapshotGRPCExposeHTTP1,
+			create: proxycfgtest.TestConfigSnapshotGRPCExposeHTTP1,
 		},
 		{
 			name:   "mesh-gateway",
-			create: proxycfg.TestConfigSnapshotMeshGateway,
+			create: proxycfgtest.TestConfigSnapshotMeshGateway,
 			setup:  nil,
 		},
 		{
 			name:   "mesh-gateway-using-federation-states",
-			create: proxycfg.TestConfigSnapshotMeshGatewayUsingFederationStates,
+			create: proxycfgtest.TestConfigSnapshotMeshGatewayUsingFederationStates,
 			setup:  nil,
 		},
 		{
 			name:   "mesh-gateway-no-services",
-			create: proxycfg.TestConfigSnapshotMeshGatewayNoServices,
+			create: proxycfgtest.TestConfigSnapshotMeshGatewayNoServices,
 			setup:  nil,
 		},
 		{
 			name:   "mesh-gateway-service-subsets",
-			create: proxycfg.TestConfigSnapshotMeshGateway,
+			create: proxycfgtest.TestConfigSnapshotMeshGateway,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.MeshGateway.ServiceResolvers = map[structs.ServiceName]*structs.ServiceResolverConfigEntry{
 					structs.NewServiceName("bar", nil): {
@@ -297,7 +298,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "mesh-gateway-ignore-extra-resolvers",
-			create: proxycfg.TestConfigSnapshotMeshGateway,
+			create: proxycfgtest.TestConfigSnapshotMeshGateway,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.MeshGateway.ServiceResolvers = map[structs.ServiceName]*structs.ServiceResolverConfigEntry{
 					structs.NewServiceName("bar", nil): {
@@ -333,7 +334,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "mesh-gateway-service-timeouts",
-			create: proxycfg.TestConfigSnapshotMeshGateway,
+			create: proxycfgtest.TestConfigSnapshotMeshGateway,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.MeshGateway.ServiceResolvers = map[structs.ServiceName]*structs.ServiceResolverConfigEntry{
 					structs.NewServiceName("bar", nil): {
@@ -355,7 +356,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "mesh-gateway-non-hash-lb-injected",
-			create: proxycfg.TestConfigSnapshotMeshGateway,
+			create: proxycfgtest.TestConfigSnapshotMeshGateway,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.MeshGateway.ServiceResolvers = map[structs.ServiceName]*structs.ServiceResolverConfigEntry{
 					structs.NewServiceName("bar", nil): {
@@ -382,7 +383,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "mesh-gateway-hash-lb-ignored",
-			create: proxycfg.TestConfigSnapshotMeshGateway,
+			create: proxycfgtest.TestConfigSnapshotMeshGateway,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.MeshGateway.ServiceResolvers = map[structs.ServiceName]*structs.ServiceResolverConfigEntry{
 					structs.NewServiceName("bar", nil): {
@@ -410,97 +411,97 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "ingress-gateway",
-			create: proxycfg.TestConfigSnapshotIngressGateway,
+			create: proxycfgtest.TestConfigSnapshotIngressGateway,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-gateway-no-services",
-			create: proxycfg.TestConfigSnapshotIngressGatewayNoServices,
+			create: proxycfgtest.TestConfigSnapshotIngressGatewayNoServices,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-with-chain",
-			create: proxycfg.TestConfigSnapshotIngress,
+			create: proxycfgtest.TestConfigSnapshotIngress,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-with-chain-external-sni",
-			create: proxycfg.TestConfigSnapshotIngressExternalSNI,
+			create: proxycfgtest.TestConfigSnapshotIngressExternalSNI,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-with-chain-and-overrides",
-			create: proxycfg.TestConfigSnapshotIngressWithOverrides,
+			create: proxycfgtest.TestConfigSnapshotIngressWithOverrides,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-with-chain-and-failover",
-			create: proxycfg.TestConfigSnapshotIngressWithFailover,
+			create: proxycfgtest.TestConfigSnapshotIngressWithFailover,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-with-tcp-chain-failover-through-remote-gateway",
-			create: proxycfg.TestConfigSnapshotIngressWithFailoverThroughRemoteGateway,
+			create: proxycfgtest.TestConfigSnapshotIngressWithFailoverThroughRemoteGateway,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-with-tcp-chain-failover-through-remote-gateway-triggered",
-			create: proxycfg.TestConfigSnapshotIngressWithFailoverThroughRemoteGatewayTriggered,
+			create: proxycfgtest.TestConfigSnapshotIngressWithFailoverThroughRemoteGatewayTriggered,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-with-tcp-chain-double-failover-through-remote-gateway",
-			create: proxycfg.TestConfigSnapshotIngressWithDoubleFailoverThroughRemoteGateway,
+			create: proxycfgtest.TestConfigSnapshotIngressWithDoubleFailoverThroughRemoteGateway,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-with-tcp-chain-double-failover-through-remote-gateway-triggered",
-			create: proxycfg.TestConfigSnapshotIngressWithDoubleFailoverThroughRemoteGatewayTriggered,
+			create: proxycfgtest.TestConfigSnapshotIngressWithDoubleFailoverThroughRemoteGatewayTriggered,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-with-tcp-chain-failover-through-local-gateway",
-			create: proxycfg.TestConfigSnapshotIngressWithFailoverThroughLocalGateway,
+			create: proxycfgtest.TestConfigSnapshotIngressWithFailoverThroughLocalGateway,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-with-tcp-chain-failover-through-local-gateway-triggered",
-			create: proxycfg.TestConfigSnapshotIngressWithFailoverThroughLocalGatewayTriggered,
+			create: proxycfgtest.TestConfigSnapshotIngressWithFailoverThroughLocalGatewayTriggered,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-with-tcp-chain-double-failover-through-local-gateway",
-			create: proxycfg.TestConfigSnapshotIngressWithDoubleFailoverThroughLocalGateway,
+			create: proxycfgtest.TestConfigSnapshotIngressWithDoubleFailoverThroughLocalGateway,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-with-tcp-chain-double-failover-through-local-gateway-triggered",
-			create: proxycfg.TestConfigSnapshotIngressWithDoubleFailoverThroughLocalGatewayTriggered,
+			create: proxycfgtest.TestConfigSnapshotIngressWithDoubleFailoverThroughLocalGatewayTriggered,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-splitter-with-resolver-redirect",
-			create: proxycfg.TestConfigSnapshotIngress_SplitterWithResolverRedirectMultiDC,
+			create: proxycfgtest.TestConfigSnapshotIngress_SplitterWithResolverRedirectMultiDC,
 			setup:  nil,
 		},
 		{
 			name:   "ingress-lb-in-resolver",
-			create: proxycfg.TestConfigSnapshotIngressWithLB,
+			create: proxycfgtest.TestConfigSnapshotIngressWithLB,
 			setup:  nil,
 		},
 		{
 			name:   "terminating-gateway",
-			create: proxycfg.TestConfigSnapshotTerminatingGateway,
+			create: proxycfgtest.TestConfigSnapshotTerminatingGateway,
 			setup:  nil,
 		},
 		{
 			name:   "terminating-gateway-no-services",
-			create: proxycfg.TestConfigSnapshotTerminatingGatewayNoServices,
+			create: proxycfgtest.TestConfigSnapshotTerminatingGatewayNoServices,
 			setup:  nil,
 		},
 		{
 			name:   "terminating-gateway-service-subsets",
-			create: proxycfg.TestConfigSnapshotTerminatingGateway,
+			create: proxycfgtest.TestConfigSnapshotTerminatingGateway,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.TerminatingGateway.ServiceResolvers = map[structs.ServiceName]*structs.ServiceResolverConfigEntry{
 					structs.NewServiceName("web", nil): {
@@ -536,7 +537,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "terminating-gateway-hostname-service-subsets",
-			create: proxycfg.TestConfigSnapshotTerminatingGateway,
+			create: proxycfgtest.TestConfigSnapshotTerminatingGateway,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.TerminatingGateway.ServiceResolvers = map[structs.ServiceName]*structs.ServiceResolverConfigEntry{
 					structs.NewServiceName("api", nil): {
@@ -568,7 +569,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "terminating-gateway-ignore-extra-resolvers",
-			create: proxycfg.TestConfigSnapshotTerminatingGateway,
+			create: proxycfgtest.TestConfigSnapshotTerminatingGateway,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.TerminatingGateway.ServiceResolvers = map[structs.ServiceName]*structs.ServiceResolverConfigEntry{
 					structs.NewServiceName("web", nil): {
@@ -607,7 +608,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "terminating-gateway-lb-config",
-			create: proxycfg.TestConfigSnapshotTerminatingGateway,
+			create: proxycfgtest.TestConfigSnapshotTerminatingGateway,
 			setup: func(snap *proxycfg.ConfigSnapshot) {
 				snap.TerminatingGateway.ServiceResolvers = map[structs.ServiceName]*structs.ServiceResolverConfigEntry{
 					structs.NewServiceName("web", nil): {
@@ -639,7 +640,7 @@ func TestClustersFromSnapshot(t *testing.T) {
 		},
 		{
 			name:   "ingress-multiple-listeners-duplicate-service",
-			create: proxycfg.TestConfigSnapshotIngress_MultipleListenersDuplicateService,
+			create: proxycfgtest.TestConfigSnapshotIngress_MultipleListenersDuplicateService,
 			setup:  nil,
 		},
 	}
